SourceCode_GPU/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dlib==19.7.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
imutils==0.5.2
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==6.0.0
pystray==0.14.4
requests==2.22.0
scipy==1.3.0
six==1.12.0
urllib3==1.25.3
virtualenv==16.6.1
moviepy==19.2.1